requirements.txt
aiogram==3.3.0
openai==1.3.8
aiohttp==3.9.1
aiosqlite==0.19.0
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dateutil==2.8.2
schedule==1.2.0
requests==2.31.0
lxml==4.9.3
cryptography>=41.0.0
python-dotenv==1.0.0